# src/backend-core/main.py
Based on the provided architecture, OpenAPI specification, and database schema, I will implement the 'backend-core' module using Python and FastAPI.

Here is the implementation:

```json
{
  "module_name": "backend-core",
  "files": {
    "main.py": "
from fastapi import FastAPI, Depends, HTTPException
from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
from pydantic import BaseModel
from sqlalchemy import create_engine, Column, Integer, String, Boolean, ForeignKey
from sqlalchemy.ext.declarative import declarative_base
from sqlalchemy.orm import sessionmaker, relationship
from passlib.context import CryptContext
from datetime import datetime, timedelta
import jwt

# Database connection
SQLALCHEMY_DATABASE_URL = \"postgresql://user:password@host:port/dbname\"
engine = create_engine(SQLALCHEMY_DATABASE_URL)
SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)

Base = declarative_base()

# Password hashing
pwd_context = CryptContext(schemes=[\"bcrypt\"], default=\"bcrypt\")

# OAuth2 scheme
oauth2_scheme = OAuth2PasswordBearer(tokenUrl=\"login\")

# Models
class User(Base):
    __tablename__ = \"users\"
    id = Column(Integer, primary_key=True)
    email = Column(String, unique=True, nullable=False)
    password_hash = Column(String, nullable=False)
    tasks = relationship(\"Task\", back_populates=\"user\")

class Task(Base):
    __tablename__ = \"tasks\"
    id = Column(Integer, primary_key=True)
    title = Column(String, nullable=False)
    description = Column(String, nullable=False)
    completed = Column(Boolean, default=False)
    user_id = Column(Integer, ForeignKey(\"users.id\"), nullable=False)
    user = relationship(\"User\", back_populates=\"tasks\")

# Schemas
class UserCreate(BaseModel):
    email: str
    password: str

class TaskCreate(BaseModel):
    title: str
    description: str

# API
app = FastAPI()

@app.post(\"/register\")
async def register_user(user: UserCreate):
    db = SessionLocal()
    if db.query(User).filter(User.email == user.email).first():
        raise HTTPException(status_code=400, detail=\"Email already exists\")
    hashed_password = pwd_context.hash(user.password)
    new_user = User(email=user.email, password_hash=hashed_password)
    db.add(new_user)
    db.commit()
    db.refresh(new_user)
    return {\"message\": \"User created successfully\"}

@app.post(\"/login\")
async def login_user(form_data: OAuth2PasswordRequestForm = Depends()):
    db = SessionLocal()
    user = db.query(User).filter(User.email == form_data.username).first()
    if not user or not pwd_context.verify(form_data.password, user.password_hash):
        raise HTTPException(status_code=401, detail=\"Invalid email or password\")
    token = jwt.encode({\"sub\": user.id}, \"secret_key\", algorithm=\"HS256\", expires_delta=timedelta(minutes=30))
    return {\"access_token\": token, \"token_type\": \"bearer\"}

@app.post(\"/tasks\")
async def create_task(task: TaskCreate, user_id: int = Depends(oauth2_scheme)):
    db = SessionLocal()
    user = db.query(User).filter(User.id == user_id).first()
    if not user:
        raise HTTPException(status_code=401, detail=\"Invalid user ID\")
    new_task = Task(title=task.title, description=task.description, user=user)
    db.add(new_task)
    db.commit()
    db.refresh(new_task)
    return {\"message\": \"Task created successfully\"}

@app.get(\"/tasks\")
async def get_tasks(user_id: int = Depends(oauth2_scheme)):
    db = SessionLocal()
    tasks = db.query(Task).filter(Task.user_id == user_id).all()
    return [{\"id\": task.id, \"title\": task.title, \"description\": task.description, \"completed\": task.completed} for task in tasks]

@app.get(\"/tasks/{task_id}\")
async def get_task(task_id: int, user_id: int = Depends(oauth2_scheme)):
    db = SessionLocal()
    task = db.query(Task).filter(Task.id == task_id, Task.user_id == user_id).first()
    if not task:
        raise HTTPException(status_code=404, detail=\"Task not found\")
    return {\"id\": task.id, \"title\": task.title, \"description\": task.description, \"completed\": task.completed}

@app.patch(\"/tasks/{task_id}\")
async def update_task(task_id: int, task: TaskCreate, user_id: int = Depends(oauth2_scheme)):
    db = SessionLocal()
    existing_task = db.query(Task).filter(Task.id == task_id, Task.user_id == user_id).first()
    if not existing_task:
        raise HTTPException(status_code=404, detail=\"Task not found\")
    existing_task.title = task.title
    existing_task.description = task.description
    db.commit()
    db.refresh(existing_task)
    return {\"message\": \"Task updated successfully\"}

@app.delete(\"/tasks/{task_id}\")
async def delete_task(task_id: int, user_id: int = Depends(oauth2_scheme)):
    db = SessionLocal()
    task = db.query(Task).filter(Task.id == task_id, Task.user_id == user_id).first()
    if not task:
        raise HTTPException(status_code=404, detail=\"Task not found\")
    db.delete(task)
    db.commit()
    return {\"message\": \"Task deleted successfully\"}
",
    "models.py": "
from sqlalchemy import Column, Integer, String, Boolean, ForeignKey
from sqlalchemy.ext.declarative import declarative_base
from sqlalchemy.orm import relationship

Base = declarative_base()

class User(Base):
    __tablename__ = \"users\"
    id = Column(Integer, primary_key=True)
    email = Column(String, unique=True, nullable=False)
    password_hash = Column(String, nullable=False)
    tasks = relationship(\"Task\", back_populates=\"user\")

class Task(Base):
    __tablename__ = \"tasks\"
    id = Column(Integer, primary_key=True)
    title = Column(String, nullable=False)
    description = Column(String, nullable=False)
    completed = Column(Boolean, default=False)
    user_id = Column(Integer, ForeignKey(\"users.id\"), nullable=False)
    user = relationship(\"User\", back_populates=\"tasks\")
"
  }
}
```

This implementation provides the following endpoints:

*   `POST /register`: Creates a new user with a unique email and password.
*   `POST /login`: Authenticates a user and returns a JWT token.
*   `POST /tasks`: Creates a new task with a title and description for the authenticated user.
*   `GET /tasks`: Retrieves a list of tasks for the authenticated user.
*   `GET /tasks/{task_id}`: Retrieves a task by ID for the authenticated user.
*   `PATCH /tasks/{task_id}`: Updates a task with a new title and description for the authenticated user.
*   `DELETE /tasks/{task_id}`: Deletes a task by ID for the authenticated user.

The code uses SQLAlchemy for database interactions, Pydantic for schema validation, and PyJWT for JWT token generation. It also includes error handling and authentication using OAuth2.

Note that you should replace the `SQLALCHEMY_DATABASE_URL` variable with your actual database connection URL and the `secret_key` variable with a secure secret key for JWT token generation.